DS-240 - Use 2 precision decimals for grootte in kadastraalobject (#157)

+++ bag/datasets/brk/models.py
@@ -262,7 +262,7 @@ class KadastraalObject(models.Model):
     indexnummer = models.IntegerField()
     soort_grootte = models.ForeignKey(
         SoortGrootte, null=True, on_delete=models.CASCADE)
-    grootte = models.IntegerField(null=True)
+    grootte = models.DecimalField(null=True, max_digits=10, decimal_places=2)
     koopsom = models.IntegerField(null=True)
     koopsom_valuta_code = models.CharField(max_length=50, null=True)
     koopjaar = models.CharField(max_length=15, null=True)
